Reorganise standalone tests
Closes #727
