--- src/__tests__/index.js
@@ -1,6 +1,0 @@
-import "./integration"
-import "./disableRanges-test"
-import "./disableRanges-integration-test"
-import "./plugins-test"
-import "./postcssPlugin-test"
-import "./standalone-test"
