+++ src/__tests__/standalone-test.js
@@ -1,12 +1,7 @@
+import configBlockNoEmpty from "./fixtures/config-block-no-empty"
+import path from "path"
 import standalone from "../standalone"
 import test from "tape"
-import path from "path"
-import chalk from "chalk"
-import stringFormatter from "../formatters/stringFormatter"
-import configBlockNoEmpty from "./fixtures/config-block-no-empty"
-import configExtendingOne from "./fixtures/config-extending-one"
-import configExtendingAnotherExtend from "./fixtures/config-extending-another-extend"
-import configExtendingThreeWithOverride from "./fixtures/config-extending-three-with-override"
 
 const fixturesPath = path.join(__dirname, "fixtures")
 
@@ -63,368 +58,6 @@ test("standalone with input css", t => {
   planned += 4
 
   t.plan(planned)
-})
-
-test("standalone with extending configuration and configBasedir", t => {
-  let planned = 0
-
-  standalone({
-    code: "a {}",
-    config: configExtendingOne,
-    configBasedir: path.join(__dirname, "fixtures"),
-  }).then(({ output, results }) => {
-    t.equal(typeof output, "string")
-    t.equal(results.length, 1)
-    t.equal(results[0].warnings.length, 1)
-    t.equal(results[0].warnings[0].rule, "block-no-empty")
-  }).catch(logError)
-  planned += 4
-
-  // Recursive extending
-  standalone({
-    code: "a {}",
-    config: configExtendingAnotherExtend,
-    configBasedir: path.join(__dirname, "fixtures"),
-  }).then(({ output, results }) => {
-    t.equal(typeof output, "string")
-    t.equal(results.length, 1)
-    t.equal(results[0].warnings.length, 1)
-    t.equal(results[0].warnings[0].rule, "block-no-empty")
-  }).catch(logError)
-  planned += 4
-
-  // Extending with overrides
-  standalone({
-    code: "a {}",
-    config: configExtendingThreeWithOverride,
-    configBasedir: path.join(__dirname, "fixtures"),
-  }).then(({ results }) => {
-    t.equal(results[0].warnings.length, 0)
-  }).catch(logError)
-  planned += 1
-
-  t.plan(planned)
-})
-
-test("standalone with extending configuration and no configBasedir", t => {
-  let planned = 0
-
-  standalone({
-    code: "a {}",
-    config: configExtendingOne,
-  }).catch(err => {
-    t.equal(err.code, 78)
-  })
-  planned += 1
-
-  t.plan(planned)
-})
-
-test("standalone with input css and alternate formatter specified by keyword", t => {
-  let planned = 0
-
-  standalone({ code: "a {}", config: configBlockNoEmpty, formatter: "string" }).then(({ output }) => {
-    const strippedOutput = chalk.stripColor(output)
-    t.equal(typeof output, "string")
-    t.ok(strippedOutput.indexOf("1:3") !== -1)
-    t.ok(strippedOutput.indexOf("block-no-empty") !== -1)
-  }).catch(logError)
-  planned += 3
-
-  t.plan(planned)
-})
-
-test("standalone with input css and alternate formatter function", t => {
-  let planned = 0
-
-  standalone({ code: "a {}", config: configBlockNoEmpty, formatter: stringFormatter }).then(({ output }) => {
-    const strippedOutput = chalk.stripColor(output)
-    t.equal(typeof output, "string")
-    t.ok(strippedOutput.indexOf("1:3") !== -1)
-    t.ok(strippedOutput.indexOf("block-no-empty") !== -1)
-  }).catch(logError)
-  planned += 3
-
-  t.plan(planned)
-})
-
-test("standalone with input css and quiet mode", t => {
-  let planned = 0
-  const config = {
-    quiet: true,
-    rules: {
-      "block-no-empty": [ true, { "severity": "warning" } ],
-    },
-  }
-
-  standalone({ code: "a {}", config }).then(({ results }) => {
-    t.deepEqual(results[0].warnings, [])
-  }).catch(logError)
-  planned += 1
-
-  t.plan(planned)
-})
-
-test("standalone with scss syntax", t => {
-  let planned = 0
-  const config = {
-    rules: {
-      "block-no-empty": true,
-    },
-  }
-
-  standalone({
-    config,
-    code: "$foo: bar; // foo;\nb {}",
-    syntax: "scss",
-    formatter: stringFormatter,
-  }).then(({ output }) => {
-    const strippedOutput = chalk.stripColor(output)
-    t.equal(typeof output, "string")
-    t.ok(strippedOutput.indexOf("2:3") !== -1)
-    t.ok(strippedOutput.indexOf("block-no-empty") !== -1)
-  }).catch(logError)
-  planned += 3
-
-  t.plan(planned)
-})
-
-test("standalone with sugarss syntax", t => {
-  let planned = 0
-  const config = {
-    rules: {
-      "length-zero-no-unit": true,
-    },
-  }
-
-  standalone({
-    config,
-    code: ".one\n  color: black\n  top: 0px\n.two",
-    syntax: "sugarss",
-    formatter: stringFormatter,
-  }).then(({ output }) => {
-    const strippedOutput = chalk.stripColor(output)
-    t.equal(typeof output, "string")
-    t.ok(strippedOutput.indexOf("3:9") !== -1)
-    t.ok(strippedOutput.indexOf("length-zero-no-unit") !== -1)
-  }).catch(logError)
-  planned += 3
-
-  t.plan(planned)
-})
-
-test("standalone with Less syntax", t => {
-  let planned = 0
-  const config = {
-    rules: {
-      "block-no-empty": true,
-    },
-  }
-
-  standalone({
-    config,
-    code: "@foo: bar; // foo;\nb {}",
-    syntax: "less",
-    formatter: stringFormatter,
-  }).then(({ output }) => {
-    const strippedOutput = chalk.stripColor(output)
-    t.equal(typeof output, "string")
-    t.ok(strippedOutput.indexOf("2:3") !== -1)
-    t.ok(strippedOutput.indexOf("block-no-empty") !== -1)
-  }).catch(logError)
-  planned += 3
-
-  t.plan(planned)
-})
-
-test("standalone with extending config and ignoreFiles glob ignoring single glob", t => {
-  standalone({
-    files: [`${fixturesPath}/*.css`],
-    config: {
-      ignoreFiles: "**/invalid-hex.css",
-      extends: [
-        "./config-block-no-empty",
-        "./config-color-no-invalid-hex",
-      ],
-    },
-    configBasedir: path.join(__dirname, "fixtures"),
-  }).then(({ results }) => {
-    t.equal(results.length, 2)
-    t.ok(results[0].source.indexOf("empty-block.css") !== -1)
-    t.equal(results[0].warnings.length, 1)
-    t.ok(results[1].source.indexOf("invalid-hex.css") !== -1)
-    t.equal(results[1].warnings.length, 0)
-    t.ok(results[1].ignored)
-  }).catch(logError)
-  t.plan(6)
-})
-
-test("standalone with absolute ignoreFiles glob path", t => {
-  standalone({
-    files: [ `${fixturesPath}/empty-block.css`, `${fixturesPath}/invalid-hex.css` ],
-    config: {
-      ignoreFiles: [`${fixturesPath}/empty-b*.css`],
-      rules: {
-        "block-no-empty": true,
-      },
-    },
-    configBasedir: path.join(__dirname, "fixtures"),
-  }).then(({ results }) => {
-    t.equal(results.length, 2)
-    t.equal(results[0].warnings.length, 0)
-    t.ok(results[0].ignored)
-    t.equal(results[1].warnings.length, 0)
-    t.notOk(results[1].ignored)
-  }).catch(logError)
-  t.plan(5)
-})
-
-test("standalone with extending config with ignoreFiles glob ignoring one by negation", t => {
-  standalone({
-    files: [`${fixturesPath}/*.css`],
-    config: {
-      ignoreFiles: [
-        "**/*.css",
-        "!**/invalid-hex.css",
-      ],
-      extends: [
-        `${fixturesPath}/config-block-no-empty`,
-        `${fixturesPath}/config-color-no-invalid-hex`,
-      ],
-    },
-    configBasedir: path.join(__dirname, "fixtures"),
-  }).then(({ results }) => {
-    t.equal(results.length, 2)
-    t.ok(results[0].source.indexOf("empty-block.css") !== -1)
-    t.equal(results[0].warnings.length, 0)
-    t.ok(results[0].ignored)
-    t.ok(results[1].source.indexOf("invalid-hex.css") !== -1)
-    t.equal(results[1].warnings.length, 1)
-    t.notOk(results[1].ignored)
-  }).catch(logError)
-  t.plan(7)
-})
-
-test("standalone with .stylelintignore file ignoring one file", t => {
-  standalone({
-    files: [`${fixturesPath}/*.css`],
-    config: {
-      extends: [
-        `${fixturesPath}/config-block-no-empty`,
-        `${fixturesPath}/config-color-no-invalid-hex`,
-      ],
-    },
-    configBasedir: path.join(__dirname, "fixtures/ignore_config"),
-  }).then(({ results }) => {
-    t.equal(results.length, 2)
-    t.ok(results[0].source.indexOf("empty-block.css") !== -1)
-    t.equal(results[0].warnings.length, 0)
-    t.ok(results[0].ignored)
-    t.ok(results[1].source.indexOf("invalid-hex.css") !== -1,
-      "violation is for block-no-empty, not color-no-invalid-hex")
-    t.equal(results[1].warnings.length, 1)
-    t.notOk(results[1].ignored)
-  }).catch(logError)
-  t.plan(7)
-})
-
-test("standalone with specified `ignorePath` file ignoring one file", t => {
-  standalone({
-    files: [`${fixturesPath}/empty-block.css`],
-    config: {
-      rules: {
-        "block-no-empty": true,
-      },
-    },
-    ignorePath: path.join(__dirname, "fixtures/ignore_config/foo.txt"),
-    configBasedir: path.join(__dirname, "fixtures"),
-  }).then(({ results }) => {
-    t.equal(results[0].warnings.length, 0)
-    t.ok(results[0].ignored)
-  }).catch(logError)
-  t.plan(2)
-})
-
-test("standalone extending a config that ignores files", t => {
-  let planned = 0
-  standalone({
-    files: [`${fixturesPath}/*.css`],
-    config: {
-      extends: [
-        `${fixturesPath}/config-extending-and-ignoring`,
-      ],
-    },
-    configBasedir: path.join(__dirname, "fixtures"),
-  }).then(({ results }) => {
-    t.equal(results.length, 2)
-    t.ok(results[0].source.indexOf("empty-block.css") !== -1,
-      "ignoreFiles in extended config has no effect")
-    t.equal(results[0].warnings.length, 1)
-    t.ok(results[1].source.indexOf("invalid-hex.css") !== -1)
-    t.equal(results[1].warnings.length, 0)
-  }).catch(logError)
-  planned += 5
-  t.plan(planned)
-})
-
-test("standalone extending a config that is overridden", t => {
-  standalone({
-    code: "a { b: \"c\" }",
-    config: {
-      extends: [
-        `${fixturesPath}/config-string-quotes-single`,
-      ],
-      rules: { "string-quotes": "double" },
-    },
-  }).then(({ results }) => {
-    t.equal(results[0].warnings.length, 0)
-  }).catch(logError)
-  t.plan(1)
-})
-
-test("standalone loading YAML with custom message", t => {
-  standalone({
-    code: "a { color: pink; }",
-    configFile: path.join(__dirname, "fixtures/config-color-named-custom-message.yaml"),
-  }).then(({ results }) => {
-    t.equal(results[0].warnings.length, 1)
-    t.equal(results[0].warnings[0].text, "Unacceptable")
-  }).catch(logError)
-
-  t.plan(2)
-})
-
-test("standalone using codeFilename and ignoreFiles together", t => {
-  standalone({
-    code: "a {}",
-    codeFilename: path.join(__dirname, "foo.css"),
-    config: {
-      ignoreFiles: ["**/foo.css"],
-      rules: { "block-no-empty": true },
-    },
-  }).then(({ results }) => {
-    t.equal(results[0].warnings.length, 0)
-    t.ok(results[0].ignored)
-  }).catch(logError)
-
-  t.plan(2)
-})
-
-test("standalone using codeFilename and ignoreFiles with configBasedir", t => {
-  standalone({
-    code: "a {}",
-    codeFilename: path.join(__dirname, "foo.css"),
-    config: {
-      ignoreFiles: ["foo.css"],
-      rules: { "block-no-empty": true },
-    },
-    configBasedir: __dirname,
-  }).then(({ results }) => {
-    t.equal(results[0].warnings.length, 0)
-    t.ok(results[0].ignored)
-  }).catch(logError)
-
-  t.plan(2)
 })
 
 test("standalone passing code with syntax error", t => {
@@ -485,80 +118,4 @@ test("configuration error sets errored to true", t => {
   t.plan(1)
 })
 
-function logError(err) {
-  console.log(err.stack) // eslint-disable-line no-console
-}
-
-import scss from "postcss-scss"
-import less from "postcss-less"
-import sugarss from "sugarss"
-
-test("standalone with syntax set by extension", t => {
-  let planned = 0
-
-  // First three sets of cases verify that the final test
-  // should be meaningful
-
-  standalone({
-    files: `${fixturesPath}/extension-sensitive.*`,
-    config: { rules: { "block-no-empty": true } },
-    syntax: "scss",
-  }).then(({ results }) => {
-    t.equal(results.length, 3, "correct number of files")
-    const sssResult = results.find(r => path.extname(r.source) === ".sss")
-    const lessResult = results.find(r => path.extname(r.source) === ".less")
-    const scssResult = results.find(r => path.extname(r.source) === ".scss")
-    t.equal(sssResult._postcssResult.opts.syntax, scss)
-    t.equal(lessResult._postcssResult.opts.syntax, scss)
-    t.equal(scssResult._postcssResult.opts.syntax, scss)
-  }).catch(logError)
-  planned += 4
-
-  standalone({
-    files: `${fixturesPath}/extension-sensitive.*`,
-    config: { rules: { "block-no-empty": true } },
-    syntax: "less",
-  }).then(({ results }) => {
-    t.equal(results.length, 3, "correct number of files")
-    const sssResult = results.find(r => path.extname(r.source) === ".sss")
-    const lessResult = results.find(r => path.extname(r.source) === ".less")
-    const scssResult = results.find(r => path.extname(r.source) === ".scss")
-    t.equal(sssResult._postcssResult.opts.syntax, less)
-    t.equal(lessResult._postcssResult.opts.syntax, less)
-    t.equal(scssResult._postcssResult.opts.syntax, less)
-  }).catch(logError)
-  planned += 4
-
-  standalone({
-    files: `${fixturesPath}/extension-sensitive.*`,
-    config: { rules: { "block-no-empty": true } },
-    syntax: "sugarss",
-  }).then(({ results }) => {
-    t.equal(results.length, 3, "correct number of files")
-    const sssResult = results.find(r => path.extname(r.source) === ".sss")
-    const lessResult = results.find(r => path.extname(r.source) === ".less")
-    const scssResult = results.find(r => path.extname(r.source) === ".scss")
-    t.equal(sssResult._postcssResult.opts.syntax, sugarss)
-    t.equal(lessResult._postcssResult.opts.syntax, sugarss)
-    t.equal(scssResult._postcssResult.opts.syntax, sugarss)
-  }).catch(logError)
-  planned += 4
-
-  standalone({
-    files: `${fixturesPath}/extension-sensitive.*`,
-    config: { rules: { "block-no-empty": true } },
-  }).then(({ results }) => {
-    t.equal(results.length, 3, "correct number of files")
-    const sssResult = results.find(r => path.extname(r.source) === ".sss")
-    const lessResult = results.find(r => path.extname(r.source) === ".less")
-    const scssResult = results.find(r => path.extname(r.source) === ".scss")
-    t.equal(sssResult._postcssResult.opts.syntax, sugarss, ".sss causes sugarss syntax")
-    t.equal(lessResult._postcssResult.opts.syntax, less, ".less causes Less syntax")
-    t.equal(scssResult._postcssResult.opts.syntax, scss, ".scss causes SCSS syntax")
-  }).catch(logError)
-  planned += 4
-
-  t.plan(planned)
-})
-
-function logError(err) { console.log(err.stack) } // eslint-disable-line
+function logError(err) { console.log(err.stack) } // eslint-disable-line no-console
